pyproject.toml
[project]
name = "watch-expense-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.19",
    "fastapi[standard]>=0.128.0",
    "mypy-boto3-dynamodb>=1.42.3",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib"]
